+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - Issue [#105](https://github.com/42BV/beanmapper/issues/105), **Ability to deal with @BeanSecuredProperty by delegating to a SecuredPropertyHandler**; when a field is tagged as @BeanSecuredProperty, BeanMapper will query its attached SecuredPropertyHandler. The handler will most likely be associated with a security implementation, such as Spring Security (not handled here). If no handler is present, access is granted by default.
+- Issue [#106](https://github.com/42BV/beanmapper/issues/106), **When a @BeanSecuredProperty is found without a SecuredPropertyHandler being set, throw an exception**; the absence of a SecuredPropertyHandler is by default a reason to throw an exception when @BeanSecuredProperty is used anywhere within the application. 
 
 ## [2.3.2] - 2018-03-06
 ### Fixed